Added "VillageMusic" event with 3 phases, and "MainMenuMusic" event
VillageMusic has VIllagePhase parameter with values 1, 2 and 3 for the different phases of village rebuild. Need to set this parameter as part of the Fmod Event Emitter in the dropdown menu so ensure the correct one plays at the correct time.

Will return to MainMenuMusic if it's approved to figure out how to fade it out smoothly.